autogen/pyproject.toml
[project]
name = "linj-autogen"
version = "0.1.0"
description = "LinJ/ContiText Autogen Implementation"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.3.1",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pyyaml>=6.0.3",
    "structlog>=25.5.0",
]

[project.scripts]
linj-autogen = "linj_autogen.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
